backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.6
httpx==0.25.2
beautifulsoup4==4.12.3
lxml==5.1.0
yfinance==0.2.36
python-socketio==5.11.1
websockets==12.0
neo4j==5.16.0
qdrant-client==1.7.3
openai==1.12.0
tiktoken==0.5.2
langchain==0.1.4
langchain-openai==0.0.5
tradingview-ta==3.3.0
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
celery-redbeat==2.2.0
flower==2.0.1
aiofiles==23.2.1
orjson==3.9.15
apscheduler==3.10.4
